+++ containers/nginx-storage/run-nginx-and-crond.sh
@@ -1,0 +1,6 @@
+#! /bin/sh -e
+
+# Run crond
+crond -f &
+# Run nginx
+/usr/local/openresty/bin/openresty -g "daemon off;"